LayoutTests/imported/w3c/web-platform-tests/workers/dedicated-worker-in-data-url-context.window-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: Worker

FAIL Create a dedicated worker in a data url frame assert_equals: expected "PASS" but got "Worker construction unexpectedly synchronously failed"
FAIL Create a dedicated worker in a data url dedicated worker assert_equals: expected "PASS" but got "Worker construction unexpectedly synchronously failed"
PASS Create a data url dedicated worker in a data url frame
FAIL Create a data url dedicated worker in a data url dedicated worker promise_test: Unhandled rejection with value: "ReferenceError: Can't find variable: Worker"
